Re-work populateOffences to be functional
