+++ packages/core/phase1/parse/transformSpiToAho/populateOffences.test.ts
@@ -1,12 +1,12 @@
 import { readFileSync } from "fs"
 import parseSpiResult from "../parseSpiResult"
-import PopulateOffences from "./PopulateOffences"
+import populateOffences from "./populateOffences"
 
 describe("PopulateOffences", () => {
   it("should transform SPI Offences to Hearing Outcome Offences", () => {
     const message = readFileSync("phase1/tests/fixtures/input-message-001-variations.xml", "utf-8")
     const courtResult = parseSpiResult(message).DeliverRequest.Message.ResultedCaseMessage
-    const result = new PopulateOffences(courtResult).execute()
+    const result = populateOffences(courtResult)
 
     expect(result).toBeDefined()
     expect(result).toMatchSnapshot()
@@ -15,7 +15,7 @@ describe("PopulateOffences", () => {
   it("should handle adjournment sine die results by setting the conviction date on offences", () => {
     const message = readFileSync("phase1/tests/fixtures/input-message-001-sine-die.xml", "utf-8")
     const courtResult = parseSpiResult(message).DeliverRequest.Message.ResultedCaseMessage
-    const result = new PopulateOffences(courtResult).execute()
+    const result = populateOffences(courtResult)
 
     expect(result).toBeDefined()
     expect(result.offences[0].ConvictionDate).toBeUndefined()
@@ -26,7 +26,7 @@ describe("PopulateOffences", () => {
   it("should add bail conditions for bail qualifiers in results", () => {
     const message = readFileSync("phase1/tests/fixtures/input-message-001-bail-qualifiers.xml", "utf-8")
     const courtResult = parseSpiResult(message).DeliverRequest.Message.ResultedCaseMessage
-    const result = new PopulateOffences(courtResult).execute()
+    const result = populateOffences(courtResult)
 
     expect(result).toBeDefined()
     expect(result.bailConditions).toStrictEqual(["With Electronic Tagging"])
